requirements.txt
bcrypt==3.2.0
boto3==1.22.0
botocore==1.25.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.12
click==8.0.2
colorama==0.4.4
DateTime==4.4
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==1.0.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
pipreqs==0.4.11
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.3
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==2.3.3
yarg==0.1.9
zope.interface==5.4.0
